Progress Problem Solved

+++ src/com/abstrucelogic/crypto/CryptoProgressListener.java
@@ -3,7 +3,7 @@ package com.abstrucelogic.crypto;
 public interface CryptoProgressListener {
 	
 	public void cryptoProcessStarted();
-	public void cryptoInProgress(int prog);
+	public void cryptoInProgress(float prog);
 	public void cryptoProcessComplete();
 	public void cryptoProcessError();
