+++ src/com/abstrucelogic/crypto/async/CryptoAsyncHandler.java
@@ -14,7 +14,7 @@ public class CryptoAsyncHandler implements CryptoHandler {
 	}
 
 	@Override
-	public void processStatusUpdate(CryptoProcessStatus status, int progressPer) {
+	public void processStatusUpdate(CryptoProcessStatus status, float progressPer) {
 		switch(status) {
 		case COMPLETE :
 			this.mCurCryptoConf.getListener().cryptoProcessComplete();
