+++ src/com/abstrucelogic/crypto/processor/CryptoProcessStatusListener.java
@@ -4,6 +4,6 @@ import com.abstrucelogic.crypto.constants.CryptoProcessStatus;
 
 public interface CryptoProcessStatusListener {
 	
-	public void processStatusUpdate(CryptoProcessStatus status, int progressPer);
+	public void processStatusUpdate(CryptoProcessStatus status, float progressPer);
 	
 }
